Add world data set handler to loadData
